Add docs for Launchers

+++ boot/src/test/java/io/datakernel/config/EffectiveConfigTest.java
@@ -94,7 +94,7 @@ public class EffectiveConfigTest {
 		ServerSocketSettings settings = config.get(converter, "Server.socketSettings");
 		settings = config.get(converter, "Server.socketSettings", settings);
 
-		assertEquals(10, settings.getBacklog().toInt());
+		assertEquals(10, settings.getBacklog());
 		assertEquals(10, settings.getReceiveBufferSize().toInt());
 		assertEquals(true, settings.getReuseAddress());
 	}
